Sat, Jun 4, 2022, 5:05 AM +00:00

+++ readability/readability.c
@@ -12,9 +12,9 @@ int main(void)
 {
     string t = get_string("Text: ");
     int l = count_letters(t);
-    int w = count_words(t);
+    float w = count_words(t);
     int s = count_sentances(t);
-    printf("%i letters\n%i words\n%i sentances\n", l, w, s);
+    printf("%i letters\n%f words\n%i sentances\n", l, w, s);
     float i = (0.0588 * (l / w * 100)) - (0.296 * (s / w * 100)) - 15.8;
     printf ("%f\n", i);
     if (i < 1)
